LayoutTests/http/wpt/preload/as-attribute-expected.txt

CONSOLE MESSAGE: <link rel=preload> cannot have the empty string as `as` value
CONSOLE MESSAGE: <link rel=preload> must have a valid `as` value


PASS Test preload as attribute values.
